backend/requirements.txt
python-dotenv>=1.0.0
python-pptx>=0.6.21
pypdf>=3.0.0
python-docx>=0.8.11
openpyxl>=3.0.0
chromadb>=0.4.0
ollama>=0.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
openai>=1.0.0
google-generativeai>=0.3.0
cohere>=5.0.0
voyageai>=0.3.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.0
keyring>=24.0.0
